upgrade fuzz tests

+++ fuzz/fuzz_targets/fuzz_fraction.rs
@@ -4,6 +4,6 @@ use libfuzzer_sys::fuzz_target;
 
 use zfuel::fraction::Fraction;
 
-fuzz_target!(|n: (i128, i128)| {
+fuzz_target!(|n: (i64, i64)| {
     let _ = Fraction::new(n.0, n.1);
 });
